refactored Menubar.hook -> LocalNav.hook. In the future the Menubar will need a different hook.
